pyproject.toml
[project]
name = "nexops-mcp"
version = "0.1.0"
description = "NexOps Model Context Protocol Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
    "websockets>=12.0",
    "langextract",
    "google-generativeai"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
